third_party/blink/web_tests/http/tests/sendbeacon/beacon-cross-origin-expected.txt

PingLoader dispatched to 'http://localhost:8000/sendbeacon/resources/save-beacon.php?name=cross-origin'.
Testing navigator.sendBeacon() cross-origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.sendBeacon("http://localhost:8000/sendbeacon/resources/save-beacon.php?name=cross-origin", "CrossOrigin"); is true
PASS Beacon sent successfully
PASS Content-Type: text/plain;charset=UTF-8
PASS Origin: http://127.0.0.1:8000
PASS Referer: http://127.0.0.1:8000/
PASS Request-Method: POST
PASS Length: 11
PASS Body: CrossOrigin
PASS 
PASS successfullyParsed is true

TEST COMPLETE
